+++ db/migrate/20150812142441_add_research_highlight_to_highlights.rb
@@ -1,0 +1,5 @@
+class AddResearchHighlightToHighlights < ActiveRecord::Migration
+  def change
+    add_column :highlights, :research_highlight, :boolean, null: false, default: false
+  end
+end